HUB-909: Pull from maven central
This updates the build.gradle to pull from maven central for public
builds. This has required a few changes:

Update the shadow plugin to pull from the gradle plugin portal.

Add the AWS KMS sdk. This was previously brought in through the
verify-event-emitter dependency. The latest version doesn't bring it
in.

Update the `EventEmitterConfiguration` to include a new field defined
on the abstract `Configuration` class from `verify-event-emitter`.

Update the `required_level_of_assurance` key in the test data to
`preferred_level_of_assurance`. The enum of allowed keys can be found in
`uk.gov.ida.eventemitter.EventDetailsKey`